requirements.txt
# Google ADK requirements for TomTom MCP integration
adk>=0.0.1
python-dotenv>=1.0.0

# Additional dependencies for testing and development
pytest>=7.0.0
requests>=2.28.0

# Flask API server dependencies
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
